Move convertor to specific file
