+++ src/extension.js
@@ -1,21 +1,7 @@
 const vscode = require('vscode')
 const disposables = require('./disposables')
 const config = require('./config')
-
-const map = {
-	C: 'Ĉ',
-	c: 'ĉ',
-	G: 'Ĝ',
-	g: 'ĝ',
-	H: 'Ĥ',
-	h: 'ĥ',
-	J: 'Ĵ',
-	j: 'ĵ',
-	S: 'Ŝ',
-	s: 'ŝ',
-	U: 'Ŭ',
-	u: 'ŭ'
-}
+const createConvertor = require('./convertor')
 
 /**
  * @param {vscode.ExtensionContext} context
@@ -24,21 +10,7 @@ function activate({ subscriptions }) {
 
 	disposables.forEach(disposable => subscriptions.push(disposable))
 
-	vscode.window.onDidChangeTextEditorSelection((event) => {
-		if (!config.activated) return
-
-		const position = event.selections[0].active
-		if (position.character < 2) return
-		const editor = event.textEditor
-		const range = new vscode.Range(position.line, position.character - 2, position.line, position.character)
-		const string = editor.document.getText(range)
-		const triggers = Object.keys(map)
-		if (['x', 'X'].includes(string[1]) && triggers.includes(string[0])) {
-			editor.edit(editBuilder => {
-				editBuilder.replace(range, map[string[0]])
-			})
-		}
-	})
+	createConvertor()
 
 	const command = vscode.commands.registerCommand('toggleActivation', async () => {
 		config.toggleActivation()
